refactor(workbench): extract reviewer-identity-lookup to lib + e2e smoke
- Move the pure cross-store lookup/matching orchestration out of the
  pages/api/workbench/reviewer-lookup route into lib/services/
  reviewer-identity-lookup.js. The route is now a thin auth + input + restriction
  shell that delegates; it re-exports lookupReviewerIdentity for back-compat
  (manual-reviewer.js + tests unchanged). Fixes the cross-route import smell
  (manual-reviewer importing a sibling page) and makes the core testable without
  the next-auth import chain.
- Add scripts/smoke-reviewer-lookup-dedup.mjs: READ-ONLY end-to-end smoke that
  discovers real subjects from live Dataverse and verifies the candidate adapters
  (findBy{Email,Orcid}Candidates, searchByName, findByContactId) and the full
  lookupReviewerIdentity orchestration. 18/18 pass live: a reviewer resolves to a
  confident self-match by email and by ORCID; none/candidates paths correct.

Verified: 30 dedup + 581 reviewer/identity/contact/orcid unit tests, build, lint,
check:api-routes (108), check:atlas, and the live smoke all green.
